Enum osztályok paramétereinek változtatása, képek finomhangolása.

See merge request szofttech-c-2024/group-01/blast-masters!39

+++ src/model/board/Velocity.java
@@ -4,17 +4,17 @@ public enum Velocity {
     BOMB_VEL(0),
     WALL_VEL(0),
     BOX_VEL(0),
-    PLAYER_VEL(2),
+    PLAYER_VEL(2), // jelenleg csak egész értékkel működik
     BONUS_VEL(0),
     MONSTER_VEL(1);
 
-    private final int velocity;
+    private final double velocity;
 
-    Velocity(int velocity) {
+    Velocity(double velocity) {
          this.velocity = velocity;
     }
 
-    public int getVelocity() {
+    public double getVelocity() {
         return velocity;
     }
 }
