+++ src/model/board/element/field/Wall.java
@@ -7,7 +7,7 @@ import java.awt.*;
 
 public class Wall extends Entity {
 
-    public Wall(int x, int y, int width, int height, int velocity, Image image, boolean alive, boolean visible) {
+    public Wall(int x, int y, int width, int height, double velocity, Image image, boolean alive, boolean visible) {
         super(x, y, width, height, velocity, image, alive, visible);
     }
